requirements.txt
requests~=2.32.3
fastapi~=0.115.12
pydantic~=2.11.4
pydantic-settings~=2.9.1
loguru~=0.7.3
redis~=6.1.0
SQLAlchemy~=2.0.41
PyJWT~=2.10.1